HSEARCH-3156 Move/rename tests for HSEARCH-1241 to avoid confusion with tests we're about to add
