--- orm/src/test/java/org/hibernate/search/test/embedded/polymorphism/Level3.java
@@ -1,60 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.test.embedded.polymorphism;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-
-import org.hibernate.search.annotations.ContainedIn;
-import org.hibernate.search.annotations.DocumentId;
-import org.hibernate.search.annotations.Field;
-import org.hibernate.search.annotations.Indexed;
-
-@Entity
-@Indexed
-class Level3 {
-
-	@Id
-	@DocumentId
-	@GeneratedValue
-	private Integer id;
-
-	@Field
-	private String name;
-
-	@OneToOne(fetch = FetchType.LAZY)
-	@ContainedIn
-	private Level2 level2Parent;
-
-	public Integer getId() {
-		return id;
-	}
-
-	public void setId(Integer id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public Level2 getLevel2Parent() {
-		return level2Parent;
-	}
-
-	public void setLevel2Parent(Level2 level2Parent) {
-		this.level2Parent = level2Parent;
-	}
-
-}
